[Refactor] Extract parameters
